Changed main dashboard structure slightly
