+++ src/main/java/concrete_classes/run/StudentManagementSystem.java
@@ -12,8 +12,9 @@ package concrete_classes.run;
  */
 public class StudentManagementSystem {
 
+    public static MainDashboard mainDashboard = new MainDashboard();
+    
     public static void main(String[] args) {
-        StudentManagementSystemRun program = new StudentManagementSystemRun();
-        program.run();
+        mainDashboard.run();
     }
 }
